CANS-529 - Resolving more Jest warnings (#210)
* Treat all console errors as test failures

This sets a higher standard for our unit tested components, which
will help reduce bugs in edge cases, deprecated methods, etc. The
changes needed for this to pass are an indication of such an
edge case.

Depending on the context, a component could trigger redirects over
and over again. Instead, we set it up to clear its shouldRedirect
state immediately after updating. In order to keep the spirit of
the existing tests, we handle lifecycle methods manually.

* Consider warnings as test failures as well