CANS-568 - Supervisor Dashboard (#231)
* Add a blank Supervisor Dashboard page

This adds a route for the supervisor dashboard page. For now, it
just renders a test string, and the particular URL is up for
debate as well.

* Add a CaseWorkerTable to the SupervisorDashboard

This currently has no data, but the structure of the page
is generally in place.

* Default case worker list to sort by name

This replaces the previous fake data with new, more realistic
fake data.

* Add Ruby layer for subordinates index

* Rename CaseWorkerTable to StaffTable

* Add a simple Subordinate Service

This adds a simple service to fetch subordinate staff people from
the Rails API. It also includes a dummy loading component to
trigger them.

* Pass real staff info to StaffTable

This hooks the StaffTable up with the SubordinateLoader to fetch
real subordinate staff people for the table.

* Switch to Component Library card

Due to some unfortunate conflicts between React Wood Duck and the
Component Library, some custom styles had to be added, primarily
around font-sizing issues.

* Use loading state of Component Library card

This requires lifting the SubordinateLoadingBoundary higher up and
passing the loading state down to the Card component.

* Move UserAccountService to common

* Load current user's name into Supervisor Table title

* Use SFC instead of PureComponent

For some reason we lint against this, even though SFCs may have
worse performance characteristics due to unnecessary rendering.

* Cache UserAccount information in the service

This change ensures that the UserAccountService only fetches the
current user information once. We do not expect it to change while
the user is in the application, and it will still refresh upon
page refresh. This helps pages like the Supervisor Dashboard and
Client Add page load more quickly.

* Simplify SubordinateCard title

There were questions about how to format the name, and so it was
simplified to a static string. This also helps with our loading
experience, which is a nice bonus.

* Move /supervisor to /staff

This is mostly in preparation for the individual staff caseload
pages which might look like /staff/:id

* Deduplicate staff mocks and proptypes

* Require staff list in StaffTable