CANS-946 - Toggleable Layouts for the rest of the pages (#507)
* Move CurrentUserLoadingBoundary to common

* ATTN!: Layout inversion for AssessmentAdd page

ATTN - This commit depends on a PR to component library, and was
built using a local link. Before merging, I must amend this commit
to use the version (likely 0.8.2), once it has been merged
and published.

This adds the layout inversion for the AssessmentAdd page. As a
requirement of that, it adds the ability to render two (left/right)
buttons in the Header, and uses an updated Component Library in
order to avoid a render loop.

* Use new layout on AssessmentEdit page

* Added StaffList and StaffRead pages

* Move files and rename things to achieve better Domain Driven Design

* Encapsulate user info loading into BreadCrumb

* Create Staff version of ChildProfilePage

* Convert Supervisor view of AssessmentPages

* Add changelog page

* Add Staff Changelog page

* Add SearchPages, remove unused components

* Move Header within Layout

* StaffLoadingBoundary shows layout while loading

* Add override for header height for RWD rem

* Temporarily fix alert styles until Layout has a slot for them

* Pass URL to breadcrumb

* Update component library

I had been using a local linked version, pending a published
version with my PR. This update includes that version.

* Fix breadcrumb on SubordinateChildProfilePage

* Add default for staffInfo