Add AssessmentManager
AssessmentManager is like a loading boundary, but fancier. Namely,
it chooses whether or not to load a new assessment or fetch an
existing one based on whether it is provided an assessment ID. It
also provides an `onSave` callback to the child, which can be used
to post/put updates to the assessment object (triggering an
updating loading state and eventual new data for the child).