Stabilize reassessment check (#626)
* Stabilize reassessment check

This stabilizes the check for whether an assessment is an initial
or subsequent assessment. Previously, it used a sleep and short
wait to see if a modal appeared, but this failed if the modal had
not yet loaded. Instead, we wait for the page title to show up.
Once it shows up, we KNOW that we have loaded the data, and so we
can check the title with no wait.

* Account for modal fade time