CANS-1079 - Expand All / Collapse All edge cases (#646)
* Remove extra layers from Domain

The Domain component had an unnecessary Fragment, div, and expanded
condition. Removing these has no visible affect on the application.

* Add key to Assessment based on age template

When the age template changes, we want to re-mount the assessment
card, as much of the local view state should be reset, including
expand/collapse functionality.

* Basic expand/collapse functionality

* Smart Expand/Collapse All behavior

* Avoiding cloning assessment when updating review state

Cloning the assessment causes each domain to compare as different
from React's perspective, which is terrible for performance (lots
of extra rendering). Instead, we leave the unchanged domains alone.

* Update Expand/Collapse all button when domains change

* Update acceptance targets, and small cleanup

* Shortcircuit review callback in Domain

* Fix regressoin of Expand All behavior when adding new domains

* Remove dead code

* Revert updateDomainIsReviewed

This change is not necessary, as the Domain will not call it unless
an update needs to happen. This pattern should be noted for other
operations, though, to avoid full cloning where possible.

* Disable flaky assertion

* Remove flaky assertions

* Regression test: check for assessment links before clicking it

* Test removal of assertions

* Fix regression test

* Fix regression test